pyproject.toml
[project]
name = "chaininsight"
version = "2.0.0"
description = "End-to-end supply chain demand planning: hierarchical forecasting + capacity planning + S&OP"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    # Core data
    "pandas>=2.2.0,<3.0",
    "numpy>=1.26.0,<3.0",
    "scipy>=1.12.0,<2.0",
    "scikit-learn>=1.4.0,<2.0",
    "pandera>=0.18.0,<1.0",
    # Visualization
    "matplotlib>=3.8.0,<4.0",
    "seaborn>=0.13.0,<1.0",
    "squarify>=0.4.3,<1.0",
    # Time-series forecasting
    "statsforecast>=1.7.0",
    "hierarchicalforecast>=0.4.0",
    "datasetsforecast>=0.0.8",
    "lightgbm>=4.0.0,<5.0",
    "xgboost>=2.0.0,<3.0",
    # MLOps
    "evidently>=0.4.0",
    # Web backend
    "fastapi>=0.109.0,<1.0",
    "uvicorn[standard]>=0.27.0,<1.0",
    "python-multipart>=0.0.7,<1.0",
    "websockets>=12.0,<14.0",
    # Database
    "sqlalchemy>=2.0.27,<3.0",
    # Config & logging
    "structlog>=23.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    # File monitoring
    "watchdog>=4.0.0,<6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0",
    "pytest-asyncio>=0.23.0,<1.0",
    "pytest-cov>=5.0.0,<7.0",
    "httpx>=0.27.0,<1.0",
    "hypothesis>=6.0.0",
    "ruff>=0.3.0,<1.0",
    "mypy>=1.8.0,<2.0",
    "pre-commit>=3.6.0",
]
chronos = [
    "chronos-forecasting>=1.3.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 130
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "S", "B"]
ignore = ["S101", "S603", "S607", "B008", "E402", "S311", "B905", "B007"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
